Restore and update rsx_lib_gl
