+++ Makefile.common
@@ -25,6 +25,8 @@ ifeq ($(HAVE_OPENGL), 1)
 
    SOURCES_C += $(LIBRETRO_DIR)/glsm/glsm.c \
                 $(LIBRETRO_DIR)/glsym/rglgen.c
+
+   SOURCES_CXX += $(CORE_DIR)/rsx/rsx_lib_gl.cpp
 endif
 
 INCFLAGS := -I$(CORE_DIR) \
